requirements.txt
schedule
redis
selenium
loguru
faker
pyyaml
